The transdMT executable has been updated.
The option to save the posterior likelihood and number of layers for all the iterations is added. This allows for monitoring the inversion convergeance.
The shell scripts have been modified to include this option. The results are saved to a .csv file.
The runMCMC.sh script and the text file defining the inversion parameters have been added to the scripts folder.