New options for plotting outputs of the inversion
The option to plot the MCMC statistics after the burn-in phase is added, it monitors RMS, likelihood and the number of layers.
THe option to plot the model responses faster is added: instead of computing a 2D histogram, the responses of 100 posterior models are plotted.
The option to plot the model response impedance Z is added.